requirements.txt
sounddevice>=0.4.6
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
